requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
requests>=2.28
python-dotenv>=1.0
